Rename and extract EditorChunkElementListItem
