+++ src/editor/panel/chunk/list/EditorChunkElementListItem.tsx
@@ -1,60 +1,15 @@
-import {ElementModel} from "../../../../data/model/element/ElementModel";
-import {UUID} from "../../../../data/model/UUID";
-import {BaseList} from "../../../component/BaseList";
-import {ElementTypes} from "../../../../data/model/element/ElementTypes";
-import {GroupModel} from "../../../../data/model/element/system/GroupModel";
-import {useEditorDispatcher} from "../../../reducer/EditorProvider";
-import React from "react";
+import {BaseVisibilitySlug} from "../../../component/slug/BaseVisibilitySlug";
+import {BaseFolderSlug} from "../../../component/slug/BaseFolderSlug";
 import {BaseInputSlug} from "../../../component/slug/BaseInputSlug";
 import {BaseActionSlug} from "../../../component/slug/BaseActionSlug";
 import {LinkIcon, LinkSlashIcon, TrashIcon} from "@heroicons/react/24/outline";
-import {BaseFolderSlug} from "../../../component/slug/BaseFolderSlug";
-import {BaseVisibilitySlug} from "../../../component/slug/BaseVisibilitySlug";
+import {ElementTypes} from "../../../../data/model/element/ElementTypes";
+import {ElementModel} from "../../../../data/model/element/ElementModel";
+import {useEditorDispatcher} from "../../../reducer/EditorProvider";
+import {GroupModel} from "../../../../data/model/element/system/GroupModel";
 import {JointModel} from "../../../../data/model/element/joint/JointModel";
 
-export type ChunkElementListProps = {
-    elements: {[key: string]: ElementModel};
-    selected: UUID[];
-}
-
-export function ChunkElementList(props: ChunkElementListProps) {
-    const dispatcher = useEditorDispatcher();
-    if (!dispatcher) {
-        return <></>;
-    }
-
-    const selectElement = (id: UUID) => {
-        dispatcher({
-            type: "editor_select_element",
-            payload: id,
-        });
-    }
-
-    const reorderElements = (childIds: UUID[], parentId: UUID | null) => {
-        dispatcher({
-            type: 'chunk_reorder_elements',
-            payload: {
-                childIds,
-                parentId,
-            },
-        });
-    }
-
-    return (
-        <BaseList
-            items={Object.values(props.elements)}
-            itemContent={element => <ChunkElementListItem {...element} />}
-            isItemSelected={element => props.selected.includes(element.id)}
-            isItemAnExpandedGroup={element => element.type === ElementTypes.Group && !(element as GroupModel).collapsed}
-            isParentOfItem={(child, parentId) => child.parent === parentId}
-            isItemHidden={child => child.hidden}
-            onSelect={selectElement}
-            onReorder={reorderElements}
-        />
-    );
-}
-
-function ChunkElementListItem(element: ElementModel) {
+export function EditorChunkElementListItem(element: ElementModel) {
     const dispatcher = useEditorDispatcher()
     if (!dispatcher) {
         return <></>;
